pyproject.toml
[project]
name            = "MSTROptionBacktest"
dynamic         = ["version"]
description     = "MSTROptionBacktest"
readme          = "readme.md"
requires-python = "==3.11.10"

dependencies = [
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "pandas",
    "ib_insync",
]

[dependency-groups]
dev = [
    "types-lxml==2025.3.30",
]
